license explanation
Former-commit-id: 567394e6599d7c3f8ebc5872ddcb64eabbbfe549
Former-commit-id: 357bd8ef93399aca753e52de8a976450f55a99ad

+++ examples/cpp/run_all.cpp
@@ -8,7 +8,8 @@
 #include "reader.h"
 #include "timer.h"
 
-typedef float real_t;
+//typedef float real_t;
+typedef double real_t;
 
 int main(int argc, char **argv) {
   using namespace std;
